requirements.txt
streamlit
openai
python-dotenv
scikit-learn
pandas
numpy
python-docx
PyPDF2
